+++ examples/platformer_3d.rs
@@ -203,8 +203,8 @@ fn setup_player(mut commands: Commands, asset_server: Res<AssetServer>) {
     cmd.insert(RigidBody::Dynamic);
     cmd.insert(Velocity::default());
     cmd.insert(Collider::capsule_y(0.5, 0.5));
-    cmd.insert(TnuaPlatformerBundle::new_with_config(
-        TnuaPlatformerConfig {
+    cmd.insert(TnuaPlatformerBundle {
+        config: TnuaPlatformerConfig {
             full_speed: 20.0,
             full_jump_height: 4.0,
             up: Vec3::Y,
@@ -226,7 +226,8 @@ fn setup_player(mut commands: Commands, asset_server: Res<AssetServer>) {
             tilt_offset_angacl: 500.0,
             turning_angvel: 10.0,
         },
-    ));
+        ..Default::default()
+    });
     cmd.insert(TnuaAnimatingState::<AnimationState>::default());
     cmd.insert(TnuaPlatformerAnimatingOutput::default());
     cmd.insert({